fix: Ignore state change for replaced subConns.
When a replaced subConn shuts down, the balancer receives
a READY->SHUTDOWN state change notification for that subConn.
If we process this as usual, then we decrement ready channels count
in the connectivityStateEvaluator. This will eventually lead to the
aggregated state of the balancer being reported as TRANSIENT_FAILURE.
This, in turn, infinitely blocks sending any RPC calls.